urisrc: stub for new element
urisrc is a bin that holds a source element for the specified
uri inside. It will be able to switch to a new uri and start
downloading as soon as the previous one is over.

To be used as a replacement for uridownloader in the adaptive
demuxers to allow directly downloading and pushing data
downstream without having to wait for a full fragment download