urisrc: add the new internal src element when uri property is set
When a new uri is set, remove the old src (if the protocol is different)
and add the new one (or keep the old one and replace its uri).

The recommended usage is that the user waits for EOS events on the src
pad of urisrc and sets a new uri after receiving it and then drop
the EOS.

This is done from a separate thread to avoid deadlocks when operating
on the source element from its own pad task.