dashdemux: Update download rate measuring to work with urisrc
It is essential for adaptive streams to be able to measure the
download rate of its fragments. Using urisrc the fragments are
downloaded and its chunks are directly pushed downstream. The
problem this raised is that this fragment pushing might lead
to pad blocking downstream due to full queues, for example.

The download rate measurement would have severe penalties in
this case as the downloading element's thread would be blocked
and the rates would appear lower than they really are, making
dashdemux select a lower bitrate alternative.

The solution this patch proposes is having an internal queue
to decouple the downloading thread. As we have no prior knowledge
of the fragment size this queue is also 'infinite' but it should
store not more than a single full fragment at most. A new fragment
only starts downloading once the queue goes empty again. While having
a infinite queue internally isn't a good idea it isn't a regression
from previous versions as uridownloader did the same internally
and there is no better option to measure download rate for now.