dashdemux: correctly update active stream pointer when refreshing mpd
When refreshing the MPD a new mpd client is created with new active
stream structures. The client was being correctly replaced in dashdemux
but nothing was being done about the stream pointers that are freed
when the client is replaced.

This patch fixes it by correctly replacing the streams when replacing the
client.