dashdemux: use failed count per stream
This way we can track if a download failed repeatedly on the the
same fragment for a on-demand stream and try to move to the next
fragment on the hope that it will work.

This is also useful when badly generated MPDs have less fragments
than they advert. In this case dashdemux will try to move to the
next one and, as there is none, will push EOS