urisrc: fix deadlock when handling error message from urisrc
When the urisrc's child posts an error message from a state change
made from the update_uri function it could cause the user to set
a new uri from the handler. The error message would be posted
holding the same lock that was used in the set_property call,
causing a deadlock.

Use separate locks to prevent this