tsdemux: allow program-number switching in playing
Use pad groups switching to add new pads for the newly selected
program when then remove the old ones. Pads are kept around in case
the user selects the same program again in the future.

To be able to properly switch the program's exposing state is stored
with the programs. The current exposed program is moved to old_program
while the newly selected hasn't all its pads exposed. After that
the old_program is removed

https://bugzilla.gnome.org/show_bug.cgi?id=730960