basemixer: new base class for mixers
Subclass of aggregator that does time alignment of buffers before
handing them to the real implementation just for the mixing.

Subclasses can implement the 'mix' function instead of the aggregate.
The implementation should just iterate over the pads, get the buffers
and mix them, generating an output buffer to be passed to the
gst_aggregator_finish_buffer function