basemixer: add prepare and adjust_time virtual functions
Those 2 functions are a need from videoaggregator.

The prepare can be used for subclasses do some extra operation
before the mixing code is executed.

The adjust_time can be used to change the mixing start and end times.
Start can be increased to skip some data while end can be reduced to
make it mix up to a certain point in time. This is used in videoaggregator
to mix with a constant output framerate that might be different from
the input.