adaptivedemux: improve stream_seek function
add more arguments to the function, allowing better control
of how seeking is done.

Knowing the flags and the playback direction allows proper
handling of snap seeking.

Knowing the playback direction also allows to control which
segment should be downloaded when the seek target is exactly
one of the fragment limits.

And also adds a new return parameter to inform of the final selected
seeking position.